% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{charlistopts}
\alias{charlistopts}
\title{Change list version of parameters into character vectors}
\usage{
charlistopts(x)
}
\arguments{
\item{x}{A list.}
}
\value{
A character vector
}
\description{
Turns a list of options into character vectors
}
\details{
The resulting character vector has values such as "name = value" where "name" was the named item of the list. Not expected to be used directly.
}
