% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{makeSignaltree}
\alias{makeSignaltree}
\title{Color the signals detected in the tree structure}
\usage{
makeSignaltree(
  result,
  entiretree,
  pv = 0.05,
  onlysignal = TRUE,
  fillcolor = "lightgrey",
  signalcolor = "Thistle",
  penwidth = 1,
  signalpenwidth = 1
)
}
\arguments{
\item{result}{A treescan object}

\item{entiretree}{A maketree object}

\item{pv}{Numeric for P-value. The default value is 0.05.}

\item{onlysignal}{If TRUE, only the part of the tree containing the detected signal, i.e., the sub-tree, is output. If FALSE, color the detected signals across the entire tree structure.}

\item{fillcolor}{A string for the color of undetected signals. The default is "lightgrey".}

\item{signalcolor}{A string for the color of detected signals. The default is "Thistle" (lightpink).}

\item{penwidth}{The thickness of the node line of undetected signals. The default is 1.}

\item{signalpenwidth}{The thickness of the node line of detected signals. The default is 1.}
}
\description{
Specifies the color of significant nodes in the entire tree structure created by the `maketree()` function
}
\details{
The `maketree()` function must be executed first because the result of `maketree()` is an argument.
}
