% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{makeSignaltree1}
\alias{makeSignaltree1}
\title{Color the signals detected in the tree structure1}
\usage{
makeSignaltree1(
  result = result,
  tree = tree,
  fillcolor = "lightgrey",
  signalcolor = "Thistle"
)
}
\arguments{
\item{result}{A treescan object}

\item{tree}{Tree Structure}

\item{fillcolor}{A string for the color of undetected signals. The default is "lightgrey".}

\item{signalcolor}{A string for the color of detected signals. The default is "Thistle" (lightpink).}
}
\description{
Specifies the color of significant nodes in the entire tree structure created by the `maketree()` function
}
\details{
The `maketree()` function must be executed first because the result of `maketree()` is an argument.
}
