% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{maketree}
\alias{maketree}
\title{Convert an object a `data.tree` structure}
\usage{
maketree(tree, fillcolor = "lightgrey", penwidth = 1)
}
\arguments{
\item{tree}{data frame}

\item{fillcolor}{A string for node color. The default is "lightgrey".}

\item{penwidth}{A string for node edge thickness. The default is 1.}
}
\description{
Creates a `data.tree` structure using hierarchical information from the entered tree file
}
\details{
This is necessary to visualize the tree structure.
}
