% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Poiscas}
\alias{Poiscas}
\title{Case file for Poisson example data}
\format{
A data frame with 12 rows and 3 variables:
\describe{
  \item{nodeID}{Node identifier}
  \item{cases}{Number of cases on each node}
  \item{population}{Number of populations on each node}
}
}
\usage{
Poiscas
}
\description{
A dataset containing node name and the number of adverse events
}
\keyword{datasets}
