% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Poisprm}
\alias{Poisprm}
\title{Parameter file for Poisson example data}
\format{
A character with 142 lines
}
\usage{
Poisprm
}
\description{
A character with parameter value
}
\keyword{datasets}
