% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{print.treescan}
\alias{print.treescan}
\title{Methods for treescan-class objects}
\usage{
\method{print}{treescan}(x, ...)
}
\arguments{
\item{x}{A treescan object}
}
\description{
Define the default methods for Treescan-class objects, which are the result objects from a call to `treescan()`
}
