% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PVcas}
\alias{PVcas}
\title{Case file for pharmacovigilance data}
\format{
A data frame with 20 rows and 3 variables:
\describe{
  \item{nodeID}{Node identifier}
  \item{cases}{Number of cases on each node}
  \item{controls}{Number of controls on each node}
}
}
\usage{
PVcas
}
\description{
A dataset containing node name and the number of adverse events
}
\keyword{datasets}
