% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{subin}
\alias{subin}
\title{Substitute new values into the input object}
\usage{
subin(x, tsparams)
}
\arguments{
\item{x}{A character vector of the form "name = value"}

\item{tsparams}{A character vector with arbitrary lines, currently imagined to be .ts.params}
}
\value{
The modified tsparams
}
\description{
Replace an existing value found in one object with a new one
}
\details{
For each line of x, the function: 1) finds the "name" and the "value" 2) checks to see whether the "name" exists in tsparams; if not, prints a warning but if so, replaces the existing line of tsparams with that line of x. Not expected to be used directly.
}
