% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tbss}
\alias{tbss}
\title{Run TreeScan software in OS when a prm file is NOT present}
\usage{
tbss(
  prmstatus = 0,
  location,
  filename,
  cas = NA,
  tre = NA,
  scan = 0,
  model = 0,
  conditional = 0,
  self = "n",
  prob = "1/2",
  start = "[0,0]",
  end = "[0,0]",
  period = "[0,0]",
  csv = "y",
  html = "n",
  llr = "n",
  tslocation = "c:/Program Files/TreeScan"
)
}
\arguments{
\item{prmstatus}{A string indicating the presence or absence of a parameter file}

\item{location}{A string containing the location of the directory where the case and tree files are located}

\item{filename}{Name of the parameter file you want to create}

\item{cas}{data frame}

\item{tre}{data frame}

\item{scan}{Interger for scan type (TREEONLY=0, TREETIME=1, TIMEONLY=2)}

\item{model}{Interger for probability model type (POISSON=0, BERNOULLI=1, UNIFORM=2, Not-Applicable=3)}

\item{conditional}{Interger for conditional type (UNCONDITIONAL=0, TOTALCASES=1, NODE=2, NODEANDTIME=3)}

\item{self}{A string for self-control design : unconditional Bernoulli only (y/n)}

\item{prob}{Numeric for case probability : (integer/integer)}

\item{start}{A string for start data time range : [integer, integer]}

\item{end}{A string for end data time range : [integer, integer]}

\item{period}{A string for data time range : [integer, integer]}

\item{csv}{A string indicating whether to create CSV results}

\item{html}{A string indicating whether to create HTML results}

\item{llr}{A string indicating whether to create LLR results}
}
\description{
Calls out into the OS to run TreeScan software, without the parameter file specified
}
\details{
A parameter file is created with the arguments of the function, and analysis is performed accordingly.
}
