% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Tree file for example data}
\format{
A data frame with 34 rows and 2 variables:
\describe{
  \item{child nodeID}
  \item{parent nodeID}
}
}
\usage{
tree
}
\description{
A dataset with child and parents node names
}
\keyword{datasets}
