% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{treescan}
\alias{treescan}
\title{Run TreeScan software in OS using prm file}
\usage{
treescan(
  prmlocation,
  prmfilename,
  tslocation = "c:/Program Files/TreeScan",
  tsbatchfilename = "treescan64",
  cleanup = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{prmlocation}{A string containing the location of the directory where the parameter file is located}

\item{prmfilename}{Name of the parameter file without the extension, i.e., no ".prm".}

\item{tslocation}{A string containing the location where the TreeScan software is installed. The default location is "c:/Program Files/TreeScan".}

\item{tsbatchfilename}{Name of the TreeScan batch file. Depending on the version of the installed TreeScan software, "treescan64" or "treescan32" is possible, and the default is "treescan64".}

\item{cleanup}{If TRUE, deletes any TreeScan output files from the OS.}

\item{verbose}{If TRUE, display the results in the R console as if running TreeScan in batch. This may be especially useful if you expect TreeScan to take a long time to run.}
}
\description{
Calls out into the OS to run TreeScan software, with the parameter file specified
}
\details{
The analysis is performed according to the options set in the parameter file.
}
