% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TreeTempcas}
\alias{TreeTempcas}
\title{Case file for TreeTemporal example data}
\format{
A data frame with 12 rows and 3 variables:
\describe{
  \item{nodeID}{Node identifier}
  \item{cases}{Number of cases on each node}
  \item{timeOfcases}{Time of occurrence of cases on each node}
}
}
\usage{
TreeTempcas
}
\description{
A dataset containing node name and the number of adverse events
}
\keyword{datasets}
