% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TreeTempprm}
\alias{TreeTempprm}
\title{Parameter file for TreeTemporal example data}
\format{
A character with 142 lines
}
\usage{
TreeTempprm
}
\description{
A character with parameter value
}
\keyword{datasets}
