% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{ts.options}
\alias{ts.options}
\title{Set or reset parameters to be used by TreeScan}
\usage{
ts.options(invals = NULL, reset = FALSE)
}
\arguments{
\item{invals}{A list with entries of the form "name = value", where the value should be in quotes unless it is a number. Alternatively, maybe a character vector whose entries are of the form "name = value". The "name" in either case should be a valid TreeScan parameter name.}

\item{reset}{If TRUE, restore the default parameter values}
}
\description{
Set or reset parameters to be used by TreeScan
}
\details{
`ts.options()` is intended to function like `par()` or `options()`.
}
