% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\name{ts_data}
\alias{ts_data}
\title{Transpose the data}
\usage{
ts_data(
  data,
  AE_code,
  drug_code,
  case,
  control,
  lv_sep,
  model,
  condition,
  rate,
  match_vars,
  N,
  Start,
  End
)
}
\arguments{
\item{AE_code}{Code for Adverse Event}

\item{drug_code}{Name for Drug or Vaccine etc.}

\item{model}{Interger for probability model type (POISSON=0, BERNOULLI=1)}

\item{condition}{Interger for conditional type (UNCONDITIONAL=0, TOTALCASES=1)}

\item{rate}{Numeric indicating expected probability for unconditional type}

\item{match_vars}{The variables you want to use for matching.}

\item{N}{matching ratio}
}
\description{
A function that converts data to return a tree structure and a count data structure.
}
\details{
A function that converts data to return a tree structure and a count data structure.
}
