% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.cas}
\alias{write.cas}
\title{Write a TreeScan cas (case) file}
\usage{
write.cas(x, location, filename, userownames = FALSE, sep = ",")
}
\arguments{
\item{x}{data frame}

\item{location}{A string containing the location of the directory where the file will be written}

\item{filename}{Name of the cas file to be created without the extension, i.e., no ".cas"}

\item{userownames}{If TRUE, write the row names into the file. The default is FALSE}

\item{sep}{A character string to separate the terms. The default separator is comma.}
}
\description{
Creates a file with the .cas extension and write input data frame in the OS
}
\details{
TreeScan software works with ASCII format files. `write.cas()` function creates input data in that format.
}
