% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.tre}
\alias{write.tre}
\title{Write a TreeScan tre (tree) file}
\usage{
write.tre(x, location, filename, userownames = FALSE, sep = ",")
}
\arguments{
\item{x}{data frame}

\item{location}{A string containing the location of the directory where the file will be written.}

\item{filename}{Name of the tree file to be created without the extension, i.e., no ".tre".}

\item{userownames}{If TRUE, write the row names into the file. The default is FALSE.}

\item{sep}{A character string to separate the terms. The default separator is comma.}
}
\description{
Create a file with the ".tre" extension and writes the input data frame in the OS
}
\details{
TreeScan software works with ASCII format files. `write.tre()` function creates input data in that format.
}
