% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.ts.prm}
\alias{write.ts.prm}
\title{Write a TreeScan prm (parameter) file}
\usage{
write.ts.prm(location, filename, matchout = TRUE)
}
\arguments{
\item{location}{A string containing the location of the directory where the file will be written}

\item{filename}{Name of the prm file to be created without the extension, i.e., no ".prm".}

\item{matchout}{If TRUE, the result file parameter is reset to share the filename given here.}
}
\description{
Creates the set of TreeScan parameter to a specified input data and other options in the OS
}
\details{
The TreeScan options can be reset or modified `ts.options()` and/or `ts.option.extra()`.
}
